refactor: move queryPagedPostHistory
