+++ packages/mask/src/extension/background-script/CryptoService.ts
@@ -1,11 +1,5 @@
 import { assertEnvironment, Environment } from '@dimensiondev/holoflows-kit'
 assertEnvironment(Environment.ManifestBackground)
 
-export {
-    encryptTo,
-    appendShareTarget,
-    decryptComment,
-    encryptComment,
-    steganographyEncodeImage,
-} from '../../../background/services/crypto'
+export * from '../../../background/services/crypto'
 export { getPartialSharedListOfPost } from './CryptoServices/getPartialSharedListOfPost'
